42 code example where tested with various SAST tools detecting flaws: - PyLint 6 - Bandit 3 - Pyre 1 - flake8 0 - Pysa 0
Some type hint related flaws are corrected as part of this PR

fixed wrong type hints in:
CWE-664/CWE-134/compliant01.py
CWE-664/CWE-843/compliant01.py
CWE-682/CWE-1335/01/compliant02.py
CWE-693/CWE-184/compliant01.py
CWE-693/CWE-184/noncompliant01.py

renamed CWE-664/CWE-502/compliant01.py to example as it does not represent compliance.
