+++ docs/Secure-Coding-Guide-for-Python/CWE-664/CWE-502/compliant01.py
@@ -1,74 +1,69 @@
 # SPDX-FileCopyrightText: OpenSSF project contributors
 # SPDX-License-Identifier: MIT
 """ Compliant Code Example """
-import hashlib
-import hmac
 import platform
-import pickle
-import secrets
+import json
 
 
 class Message(object):
     """Sample Message Object"""
-    sender_id = 42
-    text = "Some text"
+    sender_id = int()
+    text = str()
+
+    def __init__(self):
+        self.sender_id = 42
+        self.text = "Some text"
 
     def printout(self):
-        """prints content to stdout to demonstrate active content"""
-        print(f"Message:sender_id={self.sender_id} text={self.text}")
+        print(f"sender_id: {self.sender_id}\ntext: {self.text}")
 
 
 class Preserver(object):
     """Demonstrating deserialisation"""
-    def __init__(self, _key):
-        self._key = _key
 
-    def can(self, _message: Message) -> tuple:
+    def can(self, _message: Message) -> str:
         """Serializes a Message object.
             Parameters:
                 _message (Message): Message object
             Returns:
-                _digest (String): HMAC digest string
-                _jar (bytes): pickled jar as string
+                _jar (bytes): jar as string
         """
-        _jar = pickle.dumps(_message)
-        _digest = hmac.new(self._key, _jar, hashlib.sha256).hexdigest()
-        return _digest, _jar
+        return json.dumps(vars(_message))
 
-    def uncan(self, _expected_digest, _jar) -> Message:
+    def uncan(self, _jar) -> Message:
         """Verifies and de-serializes a Message object.
             Parameters:
-                _expected_digest (String): Message HMAC digest
-                _jar (bytes): Pickled jar
+                _jar (String): Pickled jar
             Returns:
                 (Message): Message object
         """
-        _digest = hmac.new(self._key, _jar, hashlib.sha256).hexdigest()
-        if _expected_digest != _digest:
-            raise ValueError("Integrity of jar compromised")
-        return pickle.loads(_jar)
+        j = json.loads(_jar)
+        _message = Message()
+        _message.sender_id = int(j["sender_id"])
+        _message.text = str(j["text"])
+        return _message
 
 
 # serialization of a normal package
-key = secrets.token_bytes()
-print(f"key={key}")
-p1 = Preserver(key)
+p1 = Preserver()
 message = Message()
-message.printout()
-digest, jar = p1.can(message)
+jar = p1.can(message)
+print(jar)
+print(type(json.loads(jar)))
 
 # sending or storing would happen here
-p2 = Preserver(key)
+p2 = Preserver()
 message = None
-message = p2.uncan(digest, jar)
+message = p2.uncan(jar)
 message.printout()
+print(message.sender_id)
 
 #####################
 # exploiting above code example
 #####################
 print("-" * 10)
 print("Attacker trying to read the message")
-message = pickle.loads(jar)
+print(jar)
 message.printout()
 
 print("-" * 10)
@@ -83,6 +78,6 @@ system
 (S'whoami;uptime;uname -a;ls -la /etc/shadow'
 tR."""
 print("Attacker trying to inject PAYLOAD")
-p3 = Preserver(b"dont know")
+p3 = Preserver()
 message = None
-message = p3.uncan(digest, PAYLOAD)
+message = p3.uncan(PAYLOAD)
