+++ docs/Secure-Coding-Guide-for-Python/CWE-664/CWE-843/compliant01.py
@@ -3,7 +3,7 @@
 """ Compliant Code Example """
 
 
-def shopping_bag(price: int, qty: int) -> int:
+def shopping_bag(price: int, qty: str) -> int:
     return int(price) * int(qty)
