+++ docs/Secure-Coding-Guide-for-Python/CWE-693/CWE-184/compliant01.py
@@ -5,17 +5,17 @@ import re
 import unicodedata
 import sys
 
-sys.stdout.reconfigure(encoding='UTF-8')
+sys.stdout.reconfigure(encoding="UTF-8")
 
 
 class TagFilter:
     """Input validation for human language"""
 
     def filter_string(self, input_string: str) -> str:
-        """ Normalize and validate untrusted string
+        """Normalize and validate untrusted string
 
-            Parameters:
-                input_string(string): String to validate
+        Parameters:
+            input_string(string): String to validate
         """
         # normalize
         _str = unicodedata.normalize("NFKC", input_string)
@@ -24,11 +24,12 @@ class TagFilter:
         _filtered_str = "".join(re.findall(r"[/\w<>\s-]+", _str))
         if len(_str) - len(_filtered_str) != 0:
             raise ValueError("Invalid input string")
- 
+
         # validate, only allow harmless tags
         for tag in re.findall("<[^>]*>", _str):
             if tag not in ["<b>", "<p>", "</p>"]:
                 raise ValueError("Invalid input tag")
+        return _str
 
 
 #####################
